configured pom.xml, changed package names, added springdoc
