+++ src/main/java/de/unistuttgart/overworldbackend/OverworldBackendApplication.java
@@ -1,0 +1,27 @@
+package de.unistuttgart.overworldbackend;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.PropertySource;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@SpringBootApplication
+public class OverworldBackendApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(OverworldBackendApplication.class, args);
+  }
+
+  @Bean
+  public WebMvcConfigurer corsConfigurer() {
+    return new WebMvcConfigurer() {
+      @Override
+      public void addCorsMappings(CorsRegistry registry) {
+        // allow CORS requests for all resources and HTTP methods from the frontend origin
+        registry.addMapping("/**").allowedMethods("OPTIONS", "HEAD", "GET", "PUT", "POST", "DELETE");
+      }
+    };
+  }
+}
