+++ src/main/java/de/unistuttgart/overworldbackend/repositories/OverworldRepository.java
@@ -1,11 +1,10 @@
-package com.example.overworldbackend.repositories;
+package de.unistuttgart.overworldbackend.repositories;
 
+import de.unistuttgart.overworldbackend.data.Configuration;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.example.overworldbackend.data.Configuration;
-
 @Repository
 public interface OverworldRepository extends JpaRepository<Configuration, Long> {
-    Configuration findByStaticWorldId(String staticWorldId);
+  Configuration findByStaticWorldId(String staticWorldId);
 }
