+++ src/main/resources/application.properties
@@ -6,8 +6,11 @@ spring.sql.init.mode=always
 spring.sql.init.platform=postgres
 spring.datasource.url=jdbc:postgresql://localhost:5432/postgres
 
+spring.datasource.username=postgres
+spring.datasource.password=postgres
+
 
 spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation=true
 spring.jpa.generate-ddl=true
 
-server.port=8081
+springdoc.swagger-ui.path=/swagger-ui
