--- src/main/resources/db.properties
@@ -1,2 +1,0 @@
-spring.datasource.username=postgres
-spring.datasource.password=asdf
