+++ src/test/java/de/unistuttgart/overworldbackend/OverworldBackendApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.overworldbackend;
+package de.unistuttgart.overworldbackend;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -6,8 +6,6 @@ import org.springframework.boot.test.context.SpringBootTest;
 @SpringBootTest
 class OverworldBackendApplicationTests {
 
-	@Test
-	void contextLoads() {
-	}
-
+  @Test
+  void contextLoads() {}
 }
